% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{LED}
\alias{LED}
\title{Life expectancy difference -- nonproportional hazards}
\usage{
LED(t, alpha1, alpha2, theta1, theta2)
}
\arguments{
\item{t}{time to which life expectancy is calculated}

\item{alpha1}{shape parameter for control}

\item{alpha2}{shape parameter for treatment}

\item{theta1}{scale parameter for control}

\item{theta2}{scale parameter for treatment}
}
\value{
life expectatncy difference
}
\description{
Life expectancy difference -- nonproportional hazards
}
