% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{LED1}
\alias{LED1}
\title{Life expectancy difference -- proportional hazards}
\usage{
LED1(t, alpha, theta, delta)
}
\arguments{
\item{t}{time to which life expectancy is calculated}

\item{alpha}{shape parameter for both arms}

\item{theta}{scale parameter for control}

\item{delta}{represents treatment effect; scale parameter for treatment is \code{theta + delta}}
}
\value{
life expectatncy difference
}
\description{
Life expectancy difference -- proportional hazards
}
