% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{WPH}
\alias{WPH}
\title{Weibull model fitting -- proportional hazards}
\usage{
WPH(data, t_LED)
}
\arguments{
\item{data}{a data frame with three variables: survival time (named time), event (0=censored, 1=event), arm}

\item{t_LED}{time to which LED is calculated}
}
\value{
data frame of posterior samples of model parameters as well as HR and LED
}
\description{
Weibull model fitting -- proportional hazards
}
